refactor: HeartRepository 인터페이스 적용
